+++ django_formwork/static/formwork/formwork.js
@@ -34,4 +34,7 @@ import "./widgets/multi_select.js";
 import "./widgets/combo_box.js";
 import "./widgets/date_picker.js";
 import "./widgets/input_mask.js";
+import "./widgets/input_number.js";
 import "./widgets/otp_input.js";
+import "./widgets/password_reveal.js";
+import "./widgets/validated_textarea.js";